Langchain_Extract_Financial_Info/requirements.txt
# Install dependencies:
beautifulsoup4
lxml
pandas
requests
pyyaml
pyarrow


